services/ai-orchestrator/requirements-core.txt
# Cartrita AI OS - Core Requirements (Without GPU Dependencies)
# Production-ready minimal dependencies for fast Docker builds

# ============================================
# Core AI/ML Libraries
# ============================================
openai>=1.40.0
anthropic>=0.30.0
langchain>=0.2.10
langchain-openai>=0.1.20
langchain-anthropic>=0.1.20
langchain-community>=0.2.10
langchain-core>=0.3.40
langgraph>=0.6.0
langsmith>=0.1.100

# ============================================
# Vector Databases & Search
# ============================================
pgvector>=0.3.0
redis>=5.0.0
faiss-cpu>=1.8.0
chromadb>=0.5.0

# ============================================
# Web Frameworks & APIs
# ============================================
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0

# ============================================
# Database & ORM
# ============================================
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# ============================================
# HTTP & WebSocket
# ============================================
httpx>=0.27.0
websockets>=12.0
aiohttp>=3.9.4

# ============================================
# Data Processing (Essential only)
# ============================================
pandas>=2.2.0
numpy>=1.26.0

# ============================================
# Search & Knowledge APIs
# ============================================
tavily-python>=0.4.0
wikipedia>=1.4.0
wolframalpha>=5.0.0

# ============================================
# Speech & Audio
# ============================================
deepgram-sdk>=3.0.0

# ============================================
# Observability & Monitoring
# ============================================
opentelemetry-distro>=0.47b0
opentelemetry-instrumentation-fastapi>=0.47b0
opentelemetry-instrumentation-sqlalchemy>=0.47b0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-exporter-jaeger>=1.21.0
structlog>=24.1.0
sentry-sdk>=2.10.0

# ============================================
# Security & Authentication
# ============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# ============================================
# Caching & Performance
# ============================================
aiocache>=0.12.0
cachetools>=5.3.0

# ============================================
# Development & Testing
# ============================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
typer>=0.12.0
loguru>=0.7.0
pydantic-extra-types>=2.8.0

# ============================================
# Performance & Async
# ============================================
uvloop>=0.19.0
httptools>=0.6.0